requirements.txt
Flask==3.0.3
requests==2.32.3
markdown==3.7
python-docx==1.1.2
jsonschema==4.23.0
bleach==6.1.0
